[FIX|UPDATE] Updates flow to add a responder in case if absent in request
If a body serializer is not added as a middleware, the request does not
come with a JSON responder as default in the current implementation.

This flow was chosen since the API flow was one that came about from a
JSON API as the primary use case.

This change checks for the JSON responder and adds it in case a client
does not use the body parser middleware which would end up in a no JSON
responder situation.