Adds a flow to set properties on each incoming request cached at start
Instead of setting properties like database after connecting for each request,
this change introduces a flow to collect common properties when the server is
initialized. Each of the property set/initialized at start, would be propagated
to the request prior to dispatching to the corresponding request handler function.

As an example, a database connection could be initialized in a promisified helper
function which resolves to a database and the same set as a property in the server
that is initialized once the resolution succeeds.

/*
  getDatabase()
    .then(database => {
      Server
        .create({})
        .setProperty('database', database)
        .start(9999)
     })
     .catch(error => {
       console.log(`Error connecting to database :: ${error}`);
       process.exit(1);
     });
*/